+++ doc.go
@@ -106,8 +106,11 @@
 //		exit := make(chan os.Signal)
 //		signal.Notify(exit, os.Interrupt)
 //
-//		c := frisbee.NewClient("127.0.0.1:8192", router)
-//		err := c.ConnectAsync()
+//		c, err := frisbee.NewClient("127.0.0.1:8192", router)
+//		if err != nil {
+//			panic(err)
+//		}
+//		err = c.ConnectAsync()
 //		if err != nil {
 //			panic(err)
 //		}